fix: change analytics histogram data to double instead of long
(cherry picked from commit c49f4f4c4b92c9a0c581ea5e6bd576f2f0af40ef)

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/analytics/model/EventAnalytics.java
@@ -18,4 +18,4 @@ package io.gravitee.apim.core.analytics.model;
 import java.util.List;
 import java.util.Map;
 
-public record EventAnalytics(Map<String, List<Long>> values) {}
+public record EventAnalytics(Map<String, List<Double>> values) {}
